+++ backend/connector-integration/src/connectors/razorpay/test.rs
@@ -110,6 +110,7 @@ mod tests {
                     key1: "dummy_key1".to_string().into(),
                 },
                 request: PaymentsAuthorizeData {
+                    authentication_data: None,
                     payment_method_data: PaymentMethodData::Card(Card {
                         card_number: RawCardNumber(
                             CardNumber::from_str("5123456789012346").unwrap(),
@@ -289,6 +290,7 @@ mod tests {
                     key1: "dummy_key1".to_string().into(),
                 },
                 request: PaymentsAuthorizeData {
+                    authentication_data: None,
                     payment_method_data: PaymentMethodData::Card(Card {
                         card_number: RawCardNumber(CardNumber::from_str("").unwrap_or_default()),
                         card_exp_month: "".to_string().into(),
@@ -419,6 +421,7 @@ mod tests {
                     key1: "dummy_key1".to_string().into(),
                 },
                 request: PaymentsAuthorizeData {
+                    authentication_data: None,
                     payment_method_data: PaymentMethodData::Card(Card {
                         card_number: RawCardNumber(CardNumber::from_str("123").unwrap_or_default()),
                         card_exp_month: "99".to_string().into(),
@@ -571,6 +574,7 @@ mod tests {
                     key1: "dummy_key1".to_string().into(),
                 },
                 request: PaymentsAuthorizeData {
+                    authentication_data: None,
                     payment_method_data: PaymentMethodData::Card(Card {
                         card_number: RawCardNumber(
                             CardNumber::from_str("5123450000000008").unwrap(),
@@ -891,6 +895,7 @@ mod tests {
                 key1: "dummy_key1".to_string().into(),
             },
             request: PaymentsAuthorizeData {
+                authentication_data: None,
                 payment_method_data: PaymentMethodData::Card(Card {
                     card_number: RawCardNumber(CardNumber::from_str("5123450000000008").unwrap()),
                     card_exp_month: "12".to_string().into(),
@@ -1077,6 +1082,7 @@ mod tests {
                 key1: "dummy_key1".to_string().into(),
             },
             request: PaymentsAuthorizeData {
+                authentication_data: None,
                 payment_method_data: PaymentMethodData::Card(Card {
                     card_number: RawCardNumber(CardNumber::from_str("5123450000000008").unwrap()),
                     card_exp_month: "12".to_string().into(),
@@ -1474,6 +1480,7 @@ mod tests {
                     key1: "invalid_key1".to_string().into(),
                 },
                 request: PaymentsAuthorizeData {
+                    authentication_data: None,
                     payment_method_data: PaymentMethodData::Card(Card {
                         card_number: Default::default(),
                         card_exp_month: "".to_string().into(),